3 reconfigured security, dao, adding new user

+++ src/main/java/pl/ething/repository/EthingUserRepository.java
@@ -4,18 +4,16 @@
  * and open the template in the editor.
  */
 package pl.ething.repository;
+
 import org.springframework.data.repository.PagingAndSortingRepository;
-import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-import org.springframework.security.access.annotation.Secured;
 import pl.ething.model.EthingUser;
-import pl.ething.model.EthingUser;
+
 /**
  *
  * @author Koksik
  */
 @RepositoryRestResource(collectionResourceRel = "user", path = "user")
-public interface EthingUserRepository extends PagingAndSortingRepository<EthingUser, Integer> {
-
-
+public interface EthingUserRepository extends PagingAndSortingRepository<EthingUser, Long> {
+    
 }
